# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Yappy Region Sync",
    "description": "Diff de DB (tablas/SP) y de parámetros SSM/Secrets entre regiones de origen y destino",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/db/diff": {
      "post": {
        "tags": [
          "db"
        ],
        "summary": "Api Db Diff",
        "operationId": "diff_db_object",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DbDiffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiffResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/db/schemas": {
      "get": {
        "tags": [
          "db"
        ],
        "summary": "Api Db Schemas",
        "description": "List the user schemas of an environment, to feed the schema picker.",
        "operationId": "list_db_schemas",
        "parameters": [
          {
            "name": "env",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Env"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemasResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/db/objects": {
      "get": {
        "tags": [
          "db"
        ],
        "summary": "Api Db Objects",
        "description": "List the tables or stored procedures of one schema, to feed the object picker.\n\nThe schema comes from :func:`api_db_schemas`, so it is already free of system\nschemas; the connection cost is the same one that listing schemas paid.",
        "operationId": "list_db_objects",
        "parameters": [
          {
            "name": "env",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Env"
            }
          },
          {
            "name": "schema",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Schema"
            }
          },
          {
            "name": "object_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DbObjectsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/execute/sql": {
      "post": {
        "tags": [
          "db"
        ],
        "summary": "Api Execute Sql",
        "description": "Run a SQL/DDL script against an environment.\n\nThis is the only write path: ``/api/compile`` merely generates. ``code`` is\nwhatever the user is looking at, so it may be a generated script, an edited\nversion of it, or free-form SQL pasted from anywhere.",
        "operationId": "execute_sql",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteSqlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/compile": {
      "post": {
        "tags": [
          "db"
        ],
        "summary": "Api Compile",
        "description": "Generate the script that would take an object from source into destination.\n\nNothing is written here: both environments are opened read-only to compare\nthem and the script comes back in the response. The caller decides whether to\nsend it to ``/api/execute/sql``, possibly after editing it.\n\nA stored procedure is recreated wholesale from the source definition\n(``CREATE OR REPLACE``). A table is not overwritten: only the columns and\nindexes that differ are emitted, as ``ALTER TABLE``, so destination data and\nits identity survive.\n\n``code_b`` (source) and ``code_a`` (destination) come back for every branch\nso the caller can seed the editor even when there is no script to apply:\n``script`` stays ``None`` when both sides are equal and ``\"\"`` when the DDL\ntext differs but the structure does not.",
        "operationId": "compile_db_object",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/query": {
      "post": {
        "tags": [
          "db"
        ],
        "summary": "Api Query",
        "description": "Run one read-only statement and return its rows.",
        "operationId": "query_db",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/migrate": {
      "post": {
        "tags": [
          "db"
        ],
        "summary": "Api Migrate",
        "description": "Migrate every table involved in a query, honouring its joins and filters.\n\nThe query is expanded into one SELECT per table; each one's rows are copied\nwith ``REPLACE INTO``. With ``dry_run`` nothing is written: the response only\nreports how many rows each table would contribute.",
        "operationId": "migrate_db_data",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MigrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MigrationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompileRequest": {
        "properties": {
          "env_b": {
            "type": "string",
            "title": "Env B"
          },
          "env_a": {
            "type": "string",
            "title": "Env A"
          },
          "object_type": {
            "type": "string",
            "title": "Object Type"
          },
          "schema_name": {
            "type": "string",
            "title": "Schema Name"
          },
          "object_name": {
            "type": "string",
            "title": "Object Name"
          }
        },
        "type": "object",
        "required": [
          "env_b",
          "env_a",
          "object_type",
          "schema_name",
          "object_name"
        ],
        "title": "CompileRequest",
        "description": "Compile an object from one environment into another (origen -> destino)."
      },
      "CompileResponse": {
        "properties": {
          "env_b": {
            "type": "string",
            "title": "Env B"
          },
          "env_a": {
            "type": "string",
            "title": "Env A"
          },
          "object_type": {
            "type": "string",
            "title": "Object Type"
          },
          "schema_name": {
            "type": "string",
            "title": "Schema Name"
          },
          "object_name": {
            "type": "string",
            "title": "Object Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "code_a": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code A"
          },
          "code_b": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code B"
          },
          "script": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Script"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "env_b",
          "env_a",
          "object_type",
          "schema_name",
          "object_name",
          "status"
        ],
        "title": "CompileResponse",
        "description": "The generated script. Compiling never writes: the caller decides whether\nand when to send ``script`` to ``/api/execute/sql``."
      },
...
      "DbDiffRequest": {
        "properties": {
          "env_a": {
            "type": "string",
            "title": "Env A"
          },
          "env_b": {
            "type": "string",
            "title": "Env B"
          },
          "schema_name": {
            "type": "string",
            "title": "Schema Name"
          },
          "object_type": {
            "type": "string",
            "title": "Object Type"
          },
          "object_name": {
            "type": "string",
            "title": "Object Name"
          },
          "include_deletes": {
            "type": "boolean",
            "title": "Include Deletes",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "env_a",
          "env_b",
          "schema_name",
          "object_type",
          "object_name"
        ],
        "title": "DbDiffRequest"
      },
      "DbObjectsResponse": {
        "properties": {
          "env": {
            "type": "string",
            "title": "Env"
          },
          "schema_name": {
            "type": "string",
            "title": "Schema Name"
          },
          "object_type": {
            "type": "string",
            "title": "Object Type"
          },
          "objects": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Objects",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "env",
          "schema_name",
          "object_type"
        ],
        "title": "DbObjectsResponse",
        "description": "The tables or stored procedures of one schema, to feed the object picker.\n\nNamed ``schema_name`` like every other model of this contract: ``schema``\nwould shadow the deprecated ``BaseModel.schema`` and warn on import."
      },
...
      "DiffResponse": {
        "properties": {
          "env_a": {
            "type": "string",
            "title": "Env A"
          },
          "env_b": {
            "type": "string",
            "title": "Env B"
          },
          "object_type": {
            "type": "string",
            "title": "Object Type"
          },
          "schema_name": {
            "type": "string",
            "title": "Schema Name"
          },
          "object_name": {
            "type": "string",
            "title": "Object Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "code_a": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code A"
          },
          "code_b": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code B"
          },
          "script": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Script"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "env_a",
          "env_b",
          "object_type",
          "schema_name",
          "object_name",
          "status"
        ],
        "title": "DiffResponse"
      },
...
      "ExecuteRequest": {
        "properties": {
          "env": {
            "type": "string",
            "title": "Env"
          },
          "object_type": {
            "type": "string",
            "title": "Object Type"
          },
          "schema_name": {
            "type": "string",
            "title": "Schema Name",
            "default": ""
          },
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "env",
          "object_type",
          "code"
        ],
        "title": "ExecuteRequest"
      },
      "ExecuteSqlResponse": {
        "properties": {
          "env": {
            "type": "string",
            "title": "Env"
          },
          "object_type": {
            "type": "string",
            "title": "Object Type"
          },
          "schema_name": {
            "type": "string",
            "title": "Schema Name"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/StatementResultInfo"
            },
            "type": "array",
            "title": "Results",
            "default": []
          },
          "ok_count": {
            "type": "integer",
            "title": "Ok Count"
          },
          "err_count": {
            "type": "integer",
            "title": "Err Count"
          }
        },
        "type": "object",
        "required": [
          "env",
          "object_type",
          "schema_name",
          "ok_count",
          "err_count"
        ],
        "title": "ExecuteSqlResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QueryRequest": {
        "properties": {
          "env": {
            "type": "string",
            "title": "Env"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 500
          }
        },
        "type": "object",
        "required": [
          "env",
          "code"
        ],
        "title": "QueryRequest"
      },
...
      "SchemasResponse": {
        "properties": {
          "env": {
            "type": "string",
            "title": "Env"
          },
          "schemas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Schemas",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "env"
        ],
        "title": "SchemasResponse"
      },
...
    }
  }
}